+++ install.py
@@ -7,7 +7,7 @@
 
 from weecfg.extension import ExtensionInstaller
 
-VERSION = '0.1.2'
+VERSION = '0.2.0'
 
 
 def loader():